move grounding logic to a dedicated plugin
